Lili/requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
anyio==4.13.0
av==17.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.16
ctranslate2==4.7.1
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.11
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
Jinja2==3.1.6
llvmlite==0.43.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.60.0
numpy==1.26.4
openai-whisper==20250625
packaging==26.0
pefile==2024.8.26
pluggy==1.6.0
protobuf==7.34.1
pycparser==3.0
Pygments==2.20.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
pypiwin32==223
pyreadline3==3.5.4
PySide6==6.7.3
PySide6_Addons==6.7.3
PySide6_Essentials==6.7.3
pytest==9.0.2
pyttsx3==2.99
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.1
rich==14.3.3
shellingham==1.5.4
shiboken6==6.7.3
sounddevice==0.5.5
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.4.1+cu124
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
